fix: break circular type import in hooks dispatcher/adapter split
Enabling the missing .dep-cruiser.awm.js sensor config (never seeded in
this environment) surfaced 6 real no-circular violations from Task 7's
hooks refactor: install.ts/uninstall.ts/status.ts (dispatchers) import
the adapter functions from claude.ts/codex.ts, while those adapters
type-only-imported InstallOptions/InstallResult/UninstallResult/
HookStatus back from the dispatcher files — a genuine dispatcher <->
adapter cycle, even though type-only imports erase at compile time and
never manifested as a runtime bug (912/912 tests passed throughout).

Hoists the five shared types into shared.ts (a dependency-free leaf
module already holding CheckResult), which both the dispatchers and the
adapters now import from instead of cross-importing each other's types.
No behavior change; dispatchers still re-export the types from their
original module paths for any external caller.

Found during post-implementation-qa on the full codex-cli-provider-plan
branch, after seeding the sensor config this environment was missing.
